+++ partiql-lang/src/main/kotlin/org/partiql/lang/eval/physical/PhysicalPlanCompilerImpl.kt
@@ -279,6 +279,7 @@ internal class PhysicalPlanCompilerImpl(
             is PartiqlPhysical.Expr.BindingsToValues -> compileBindingsToValues(expr)
             is PartiqlPhysical.Expr.Pivot -> compilePivot(expr, metas)
             is PartiqlPhysical.Expr.GraphMatch -> TODO("Physical compilation of GraphMatch expression")
+            is PartiqlPhysical.Expr.Timestamp -> TODO()
         }
     }