+++ partiql-lang/src/main/kotlin/org/partiql/lang/prettyprint/ASTPrettyPrinter.kt
@@ -577,6 +577,7 @@ class ASTPrettyPrinter {
             )
             is PartiqlAst.Expr.GraphMatch -> TODO("Unsupported GraphMatch AST node")
             is PartiqlAst.Expr.CallWindow -> TODO("PrettyPrinter doesn't support Window Function yet.")
+            is PartiqlAst.Expr.Timestamp -> TODO()
         }
 
     private fun toRecursionTreeList(nodes: List<PartiqlAst.Expr>, attrOfParent: String? = null): List<RecursionTree> =